(goodluck) Store component data as class instances (#46)
When we use Goodluck in our games we store component data in object
literals (typed via TS interfaces), but we do that due to the code size
constaint of https://js13kgames.com. In canonical JS/TS we'd use classes
to describe the data, and class instances to store it in component
arrays.

The results are slightly better with this PR. I suspect it's due to
the fact that the engine can make more assumptions about the shape of
the objects stored in the component arrays.

@ooflorent, is this fair? If you think it is, I'll be happy to refactor
the remaining test cases.